+++ AddOns/GSE/CHANGELOG.md
@@ -1,6 +1,7 @@
 # GSE2: Gnome Sequencer Enhanced
 
-## [2.6.43](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/tree/2.6.43) (2021-01-28)
-[Full Changelog](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/compare/2.6.42...2.6.43) [Previous Releases](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/releases)
+## [2.6.44](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/tree/2.6.44) (2021-02-01)
+[Full Changelog](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/compare/2.6.43...2.6.44) [Previous Releases](https://github.com/TimothyLuke/GnomeSequencer-Enhanced/releases)
 
-- #774 Updated code to best practice.  
+- #779 Error Formatting  
+- #779 Add warning that PostMacro is in an invalid state